Исправление NumberValidator в соответствии с требованиями из pull request
