+++ cs/HomeExercises/HomeExercises.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HomeExercises", "HomeExerci
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Person.Tests", "Person.Tests\Person.Tests.csproj", "{CB57B3E5-40D3-46DC-8A72-4663207B0631}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NumberValidator.Tests", "NumberValidator.Tests\NumberValidator.Tests.csproj", "{28D5585A-51C7-4353-8F4F-F86887F0576E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{CB57B3E5-40D3-46DC-8A72-4663207B0631}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CB57B3E5-40D3-46DC-8A72-4663207B0631}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CB57B3E5-40D3-46DC-8A72-4663207B0631}.Release|Any CPU.Build.0 = Release|Any CPU
+		{28D5585A-51C7-4353-8F4F-F86887F0576E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{28D5585A-51C7-4353-8F4F-F86887F0576E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{28D5585A-51C7-4353-8F4F-F86887F0576E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{28D5585A-51C7-4353-8F4F-F86887F0576E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
